+++ cmd/stanza/init_common.go
@@ -8,6 +8,7 @@ import (
 	_ "github.com/observiq/stanza/operator/builtin/input/file"
 	_ "github.com/observiq/stanza/operator/builtin/input/forward"
 	_ "github.com/observiq/stanza/operator/builtin/input/generate"
+	_ "github.com/observiq/stanza/operator/builtin/input/goflow"
 	_ "github.com/observiq/stanza/operator/builtin/input/k8sevent"
 	_ "github.com/observiq/stanza/operator/builtin/input/stanza"
 	_ "github.com/observiq/stanza/operator/builtin/input/stdin"